pyproject.toml
[project]
name = "gnx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "cloudpickle>=3.1.1",
    "gnxlib",
    "huggingface-hub>=0.33.4",
    "jax>=0.7.2",
    "jaxlib",
    "kaleido>=1.0.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "optax>=0.2.4",
    "ott-jax>=0.5.0",
    "pillow>=11.3.0",
    "plotly>=6.1.2",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "treescope>=0.1.9",
    "wandb>=0.19.11",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.7",
    "pyright[nodejs]>=1.1.402",
    "pytest>=8.3.5",
    "pytorch-fid>=0.3.0",
    "seaborn>=0.13.2",
    "smalldiffusion>=0.4.4",
    "torch>=2.8.0",
    "torch-ema>=0.3",
    "torchvision>=0.23.0",
]
cuda = [
    "jax[cuda]>=0.7.2",
    "nvidia-cublas-cu12>=12.9.1.4",
    "nvidia-cuda-cupti-cu12>=12.9.79",
    "nvidia-cuda-nvcc-cu12>=12.9.86",
    "nvidia-cuda-nvrtc-cu12>=12.9.86",
    "nvidia-cuda-runtime-cu12>=12.9.79",
    "nvidia-cudnn-cu12>=9.14.0.64",
    "nvidia-cufft-cu12>=11.4.1.4",
    "nvidia-cusolver-cu12>=11.7.5.82",
    "nvidia-cusparse-cu12>=12.5.10.65",
    "nvidia-nccl-cu12>=2.28.3",
    "nvidia-nvjitlink-cu12>=12.9.86",
    "nvidia-nvshmem-cu12>=3.4.5",
]

[project.scripts]
gnx = "gnx.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
smalldiffusion = { git = "https://github.com/danielpfrommer-tri/smalldiffusion.git" }


jax = { git = "https://github.com/mattjj/jax.git", branch = "issue32446" }
jax_cuda12_plugin = [{ index = "jax-nightly" }]
jax_cuda12_pjrt= [{ index = "jax-nightly" }]
jaxlib = [{ index = "jax-nightly" }]
gnxlib = { workspace = true }

[tool.uv]
override-dependencies = [
   "jax-cuda12-plugin==0.8.0.dev20251010",
   "jax-cuda12-pjrt==0.8.0.dev20251010"
]


[[tool.uv.index]]
name = "jax-nightly"
url = "https://us-python.pkg.dev/ml-oss-artifacts-published/jax-public-nightly-artifacts-registry/simple/"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.workspace]
members = [
    "rust/gnxlib",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gnx"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
typeCheckingMode = "standard"
reportIncompatibleMethodOverride = false

[tool.pytest.ini_options]
markers = [
    "full: A full end-to-end training test.",
]
